First Commit
This is the first commit. Here the stm32wl55jc2 board is used for LoRa
transmission.Firstly, the FDTI module is used to take command from user
when command is received the callback function defined will send the
user input to next nucleo board.The next nucleo board listens to the
data sent by first nucleo board and will display it into the serial
communicator.